+++ app/safe/internal/state/secret/collection/read.go
@@ -11,16 +11,10 @@
 package collection
 
 import (
-	"strings"
-	"time"
-
 	"github.com/vmware-tanzu/secrets-manager/app/safe/internal/state/io"
-	"github.com/vmware-tanzu/secrets-manager/app/safe/internal/state/secret/queue/deletion"
-	"github.com/vmware-tanzu/secrets-manager/app/safe/internal/state/secret/queue/insertion"
 	"github.com/vmware-tanzu/secrets-manager/app/safe/internal/state/stats"
 	"github.com/vmware-tanzu/secrets-manager/core/crypto"
 	entity "github.com/vmware-tanzu/secrets-manager/core/entity/v1/data"
-	"github.com/vmware-tanzu/secrets-manager/core/env"
 	log "github.com/vmware-tanzu/secrets-manager/core/log/std"
 )
 
@@ -30,24 +24,24 @@ import (
 // to load the secret by name from the populated cache.
 //
 // Parameters:
-//   - cid  string: A correlation ID used to track the request and associated logging.
-//     This ID helps in tracing and debugging operations across different components
-//     or services that handle the secret data.
+//   - cid  string: A correlation ID used to track the request and associated
+//     logging. This ID helps in tracing and debugging operations across
+//     different components or services that handle the secret data.
 //   - name string: The name of the secret to be retrieved.
 //
 // Returns:
-//   - *entity.Secret: A pointer to the Secret entity if found. The Secret structure
-//     includes fields such as Name, Created, Updated, NotBefore, and ExpiresAfter.
-//     Each of these timestamp fields is converted from the stored format to a
-//     JSON compatible format. Returns nil if no secret with the provided name is
-//     found in the cache.
+//   - *entity.Secret: A pointer to the Secret entity if found. The Secret
+//     structure includes fields such as Name, Created, Updated, NotBefore, and
+//     ExpiresAfter. Each of these timestamp fields is converted from the stored
+//     format to a JSON compatible format. Returns nil if no secret with the
+//     provided name is found in the cache.
 //
 // Error Handling:
-//   - If there is an error in populating the secrets from the disk (e.g., due to
-//     read errors or  data corruption), the function logs a warning message with
-//     the correlation ID and the error message but continues execution. This does
-//     not halt the function, and it subsequently tries to fetch the secret if
-//     already available in the cache.
+//   - If there is an error in populating the secrets from the disk (e.g., due
+//     to read errors or  data corruption), the function logs a warning message
+//     with the correlation ID and the error message but continues execution.
+//     This does not halt the function, and it subsequently tries to fetch the
+//     secret if already available in the cache.
 func SecretByName(cid string, name string) *entity.Secret {
 	// Check existing stored secrets files.
 	// If VSecM pod is evicted and revived, it will not have knowledge about
@@ -55,7 +49,8 @@ func SecretByName(cid string, name string) *entity.Secret {
 	if !SecretsPopulated() {
 		err := PopulateSecrets(cid)
 		if err != nil {
-			log.WarnLn(&cid, "Failed to populate secrets from disk", err.Error())
+			log.WarnLn(&cid,
+				"Failed to populate secrets from disk", err.Error())
 		}
 	}
 
@@ -80,8 +75,9 @@ const keystoneWorkloadId = "vsecm-keystone"
 // KeystoneInitialized checks whether the keystone secret is registered.
 //
 // This is a utility function that depends on the SecretByName function to check
-// for the presence of the specific secret. A return value of true indicates that
-// the keystone is initialized and ready for use, while false indicates it is not.
+// for the presence of the specific secret. A return value of true indicates
+// that the keystone is initialized and ready for use, while false indicates it
+// is not.
 //
 // Parameters:
 //   - cid string: A correlation ID used for logging and tracing.
@@ -105,7 +101,8 @@ func AllSecrets(cid string) []entity.Secret {
 	if !SecretsPopulated() {
 		err := PopulateSecrets(cid)
 		if err != nil {
-			log.WarnLn(&cid, "Failed to populate secrets from disk", err.Error())
+			log.WarnLn(&cid,
+				"Failed to populate secrets from disk", err.Error())
 		}
 	}
 
@@ -143,7 +140,8 @@ func AllSecretsEncrypted(cid string) []entity.SecretEncrypted {
 	if !SecretsPopulated() {
 		err := PopulateSecrets(cid)
 		if err != nil {
-			log.WarnLn(&cid, "Failed to populate secrets from disk", err.Error())
+			log.WarnLn(&cid,
+				"Failed to populate secrets from disk", err.Error())
 		}
 	}
 
@@ -185,149 +183,6 @@ func contains(s []string, e string) bool {
 	return false
 }
 
-// UpsertSecret takes an entity.SecretStored object and inserts it into
-// the in-memory store if it doesn't exist, or updates it if it does. It also
-// handles updating the backing store and Kubernetes secrets if necessary.
-// If appendValue is true, the new value will be appended to the existing values,
-// otherwise it will replace the existing values.
-func UpsertSecret(secretStored entity.SecretStored, appendValue bool) {
-	cid := secretStored.Meta.CorrelationId
-
-	vs := secretStored.Values
-
-	if len(vs) == 0 {
-		log.InfoLn(&cid, "UpsertSecret: nothing to upsert. exiting.", "len(vs)", len(vs))
-		return
-	}
-
-	var nonEmptyValues []string
-	for _, value := range secretStored.Values {
-		if value != "" {
-			nonEmptyValues = append(nonEmptyValues, value)
-		}
-	}
-
-	if nonEmptyValues == nil {
-		log.InfoLn(&cid, "UpsertSecret: nothing to upsert. exiting.", "len(vs)", len(vs))
-		return
-	}
-
-	secretStored.Values = nonEmptyValues
-
-	s, exists := Secrets.Load(secretStored.Name)
-	now := time.Now()
-	if exists {
-		log.TraceLn(&cid, "UpsertSecret: Secret exists. Will update.")
-
-		ss := s.(entity.SecretStored)
-		secretStored.Created = ss.Created
-
-		if appendValue {
-			log.TraceLn(&cid, "UpsertSecret: Will append value.")
-
-			for _, v := range ss.Values {
-				if contains(secretStored.Values, v) {
-					continue
-				}
-				if len(v) == 0 {
-					continue
-				}
-				secretStored.Values = append(secretStored.Values, v)
-			}
-		}
-	} else {
-		secretStored.Created = now
-	}
-	secretStored.Updated = now
-
-	log.InfoLn(&cid, "UpsertSecret:",
-		"created", secretStored.Created, "updated", secretStored.Updated,
-		"name", secretStored.Name, "len(vs)", len(vs),
-	)
-
-	log.TraceLn(&cid, "UpsertSecret: Will parse secret.")
-
-	parsedStr, err := secretStored.Parse()
-	if err != nil {
-		log.InfoLn(&cid,
-			"UpsertSecret: Error parsing secret. Will use fallback value.",
-			err.Error(),
-		)
-	}
-
-	log.TraceLn(&cid, "UpsertSecret: Parsed secret. Will set transformed value.")
-
-	secretStored.ValueTransformed = parsedStr
-	stats.CurrentState.Increment(secretStored.Name, Secrets.Load)
-	Secrets.Store(secretStored.Name, secretStored)
-
-	// TODO: this section shall vary on the global backing store type.
-	log.TraceLn(
-		&cid, "UpsertSecret: Will push secret. len",
-		len(insertion.SecretUpsertQueue),
-		"cap", cap(insertion.SecretUpsertQueue))
-	insertion.SecretUpsertQueue <- secretStored
-	log.TraceLn(
-		&cid, "UpsertSecret: Pushed secret. len",
-		len(insertion.SecretUpsertQueue), "cap",
-		cap(insertion.SecretUpsertQueue))
-
-	// If the "name" of the secret has the prefix "k8s:", then store it as a
-	// Kubernetes secret too.
-	if strings.HasPrefix(secretStored.Name, env.StoreWorkloadAsK8sSecretPrefix()) {
-		log.TraceLn(
-			&cid,
-			"UpsertSecret: will push Kubernetes secret. len",
-			len(insertion.K8sSecretUpsertQueue),
-			"cap", cap(insertion.K8sSecretUpsertQueue),
-		)
-		insertion.K8sSecretUpsertQueue <- secretStored
-		log.TraceLn(
-			&cid,
-			"UpsertSecret: pushed Kubernetes secret. len",
-			len(insertion.K8sSecretUpsertQueue),
-			"cap", cap(insertion.K8sSecretUpsertQueue),
-		)
-	}
-}
-
-// DeleteSecret orchestrates the deletion of a specified secret from both the
-// application's internal cache and its persisted storage locations, which may
-// include local filesystem and Kubernetes secrets. The deletion process is
-// contingent upon the secret's metadata, specifically its backing store and
-// whether it is used as a Kubernetes secret.
-//
-// Parameters:
-//   - secretToDelete (entity.SecretStored): The secret entity marked for deletion,
-//     containing necessary metadata such as the name of the secret, its correlation
-//     ID for logging, and metadata specifying where and how the secret is stored.
-func DeleteSecret(secretToDelete entity.SecretStored) {
-	cid := secretToDelete.Meta.CorrelationId
-
-	_, exists := Secrets.Load(secretToDelete.Name)
-	if !exists {
-		log.WarnLn(&cid,
-			"DeleteSecret: Secret does not exist. Cannot delete.",
-			secretToDelete.Name)
-		return
-	}
-
-	// TODO: this shall switch based on the global backing store type.
-	log.TraceLn(
-		&cid, "DeleteSecret: Will delete secret. len",
-		len(deletion.SecretDeleteQueue),
-		"cap", cap(deletion.SecretDeleteQueue))
-	deletion.SecretDeleteQueue <- secretToDelete
-	log.TraceLn(
-		&cid, "DeleteSecret: Pushed secret to delete. len",
-		len(deletion.SecretDeleteQueue), "cap",
-		cap(deletion.SecretDeleteQueue))
-
-	// Remove the secret from the memory.
-	stats.CurrentState.Decrement(secretToDelete.Name, Secrets.Load)
-	Secrets.Delete(secretToDelete.Name)
-}
-
 // ReadSecret takes a key string and returns a pointer to an entity.SecretStored
 // object if the secret exists in the in-memory store. If the secret is not
 // found in memory, it attempts to read it from disk, store it in memory, and
@@ -338,7 +193,8 @@ func ReadSecret(cid string, key string) (*entity.SecretStored, error) {
 	result, secretFoundInMemory := Secrets.Load(key)
 	if secretFoundInMemory {
 		s := result.(entity.SecretStored)
-		log.TraceLn(&cid, "ReadSecret: returning from memory.", "len", len(s.Values))
+		log.TraceLn(&cid,
+			"ReadSecret: returning from memory.", "len", len(s.Values))
 		return &s, nil
 	}
 
@@ -354,6 +210,7 @@ func ReadSecret(cid string, key string) (*entity.SecretStored, error) {
 	stats.CurrentState.Increment(stored.Name, Secrets.Load)
 	Secrets.Store(stored.Name, *stored)
 
-	log.TraceLn(&cid, "ReadSecret: returning from disk.", "len", len(stored.Values))
+	log.TraceLn(&cid,
+		"ReadSecret: returning from disk.", "len", len(stored.Values))
 	return stored, nil
 }
